document that fields before cached_pid break it
Adding a field above this results in either that field or the tid field
(depending on if it's above that too) being set to zero by vfork.
Ideally that assembly would be replaced with a C++ wrapper but in the
meantime this adds a comment to save time for someone else.

Signed-off-by: anupritaisno1 <www.anuprita804@gmail.com>
(cherry picked from commit 055542a6da788b232534cc9b8f1346fe7c35122f)